requirements.txt
python-telegram-bot==13.7
notion-client==2.2.1
requests==2.26.0
python-dotenv==0.19.0
openai==1.11
pydub
ffmpeg
beautifulsoup4==4.12.3
PyMuPDF==1.24.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
pypdf==4.2.0